requirements.txt
fastapi
uvicorn
huggingface_hub
Pillow
python-dotenv